+++ BitFlow/BitFlow.py
@@ -5,6 +5,7 @@ from .Eval.NumEval import NumEval
 from .Eval.TorchEval import TorchEval
 from .Optimization import BitFlowVisitor, BitFlowOptimizer
 from .AddRoundNodes import AddRoundNodes
+from .utils import GeneratedDataset
 
 import torch
 from torch.utils import data
@@ -13,61 +14,6 @@ import random
 import math
 import copy
 import matplotlib.pyplot as plt
-
-
-class Dataset(data.Dataset):
-    def __init__(self, model, dataset_size, size_p, size_r, size_output, data_range, range_bits, true_width, dist):
-        self.X = {k: [] for k in data_range}
-        self.Y = []
-        self.data_range = data_range
-
-        P = torch.Tensor(1, size_p).fill_(true_width)[0]
-        R = torch.Tensor(1, size_r).fill_(true_width)[0]
-        torch.manual_seed(42)
-
-        for key in data_range:
-            # Create random tensor
-            input_range = data_range[key]
-
-            # calculate range bounds using range bits
-            ib = range_bits[key]
-            min_range = -1 * (2 ** (ib - 1))
-            max_range = 2 ** (ib - 1) - 1
-
-            val = 0
-            if dist == 1:
-                mean = (input_range[1]+input_range[0])/2
-                std = (mean - input_range[0])/3
-                val = torch.normal(
-                    mean=mean, std=std, size=(1, dataset_size)).squeeze()
-            elif dist == 2:
-                beta = torch.distributions.beta.Beta(
-                    torch.tensor([0.5]), torch.tensor([0.5]))
-                val = (input_range[1] - input_range[0]) * \
-                    beta.sample((dataset_size,)).squeeze() + \
-                    input_range[0]
-            else:
-                val = (input_range[1] - input_range[0]) * \
-                    torch.rand(dataset_size) + input_range[0]
-
-            val = torch.clamp(val, min_range, max_range)
-            self.X[key] = val
-
-        for i in range(dataset_size):
-            inputs = {k: self.X[k][i] for k in data_range}
-
-            inputs["P"] = P
-            inputs["R"] = R
-            inputs["O"] = torch.Tensor(
-                1, size_output).fill_(true_width)[0]
-            new_y = model(**inputs)
-            self.Y.append(new_y)
-
-    def __len__(self):
-        return len(self.X[list(self.data_range.keys())[0]])
-
-    def __getitem__(self, index):
-        return {k: self.X[k][index] for k in self.data_range}, self.Y[index]
 
 
 class BitFlow:
@@ -115,7 +61,7 @@ class BitFlow:
             (X, Y): generated data
         """
 
-        return Dataset(model, dataset_size, size_p, size_r, size_output, data_range, range_bits, true_width, dist)
+        return GeneratedDataset(model, dataset_size, size_p, size_r, size_output, data_range, range_bits, true_width, dist)
 
     def update_dag(self, dag):
         """
